requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions

django~=3.0.0
django-environ

djangorestframework
pillow 

django-storages[azure]